+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # v0.20 (unreleased)
 
+Added support for storing org-mode links from inside Helpful buffers.
+
 # v0.19 (released 12th May 2022)
 
 Fixed a hang when looking at functions that had advice but hadn't yet